refactor: useIntegrationListQuery - move useIntegrationListQuery to hooks/query
