+++ src/hooks/query/index.ts
@@ -1,8 +1,12 @@
+import { useConnectionsListQuery } from './useConnectionsListQuery';
+import { useListIntegrationsQuery } from './useIntegrationListQuery';
 import { useListInstallationsQuery } from './useListInstallationsQuery';
 import { useListProviderAppsQuery } from './useListProviderAppsQuery';
 import { useOauthConnectQuery } from './useOauthConnectQuery';
 
 export {
+  useConnectionsListQuery,
+  useListIntegrationsQuery,
   useListInstallationsQuery,
   useListProviderAppsQuery,
   useOauthConnectQuery,
